chore: save local changes before merge
